// sourcecode/BTrak.Packages/snova-dashboard-module/Dashboard-module-components/dashboard-components/src/lib/snovasys-dashboard-management/models/performanceConfigurationModel.ts
export class PerformanceConfigurationModel {
    configurationId: string;
    configurationName: string;
    formJson: string;
    timeStamp: any;
    isArchived: boolean;
    createdByUserId: string;
    createdBy: string;
    createdByImage: string;
    createdDatetime: Date;
    isDraft: boolean;
    considerRole: boolean;
    selectedRoleIds: any[];
    roleNames: string;
    ofUserId: string;
}